Move "boost" handling in term-level queries to a trait so it can be reused
