+++ src/Search/Query/TermLevel/BoostableQuery.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Nord\Lumen\Elasticsearch\Search\Query\TermLevel;
+
+use Nord\Lumen\Elasticsearch\Exceptions\InvalidArgument;
+
+/**
+ * Trait BoostableQuery
+ * @package Nord\Lumen\Elasticsearch\Search\Query\TermLevel
+ */
+trait BoostableQuery
+{
+
+    /**
+     * @var float Sets the boost value of the query, defaults to 1.0.
+     */
+    protected $boost;
+
+    /**
+     * @return float
+     */
+    public function getBoost()
+    {
+        return $this->boost;
+    }
+
+    /**
+     * @param float $boost
+     *
+     * @return $this
+     *
+     * @throws InvalidArgument
+     */
+    public function setBoost($boost)
+    {
+        $this->assertBoost($boost);
+        $this->boost = $boost;
+
+        return $this;
+    }
+
+    /**
+     * @param float $boost
+     *
+     * @throws InvalidArgument
+     */
+    protected function assertBoost($boost)
+    {
+        if (!is_float($boost)) {
+            throw new InvalidArgument(sprintf(
+                '`boost` must be a float value, "%s" given.',
+                gettype($boost)
+            ));
+        }
+    }
+}
